Suggestions for refactoring code.
I've also done a bit of tidy-up in the Keyvault code moving KeyVaultSettings into its own module.
